+++ src/main/java/org/blacksmith/finlib/interestbasis/daycount/ActActAfbConvention.java
@@ -8,15 +8,15 @@ import org.blacksmith.finlib.interestbasis.ScheduleInfo;
 public class ActActAfbConvention implements DayCountConventionCalculator {
 
   @Override
-  public int calculateDays(LocalDate startDate, LocalDate endDate, ScheduleInfo scheduleInfo) {
+  public long calculateDays(LocalDate startDate, LocalDate endDate, ScheduleInfo scheduleInfo) {
     return DateUtils.daysBetween(startDate, endDate);
   }
 
   @Override
   public double calculateYearFraction(LocalDate startDate, LocalDate endDate, ScheduleInfo scheduleInfo) {
-    final int years = DateUtils.yearsBetween(startDate, endDate);
+    final long years = DateUtils.yearsBetween(startDate, endDate);
     final LocalDate remainedPeriodEndDate = (years==0) ? endDate: endDate.minusYears(years);
-    final int remainedPeriodDays = calculateDays(startDate, remainedPeriodEndDate, scheduleInfo);
+    final long remainedPeriodDays = calculateDays(startDate, remainedPeriodEndDate, scheduleInfo);
     final double remainedPeriodYearLength = DateUtils.isLeapDayInPeriod(startDate,remainedPeriodEndDate) ? 366d : 365d;
     return years + (remainedPeriodDays / remainedPeriodYearLength);
   }
