+++ src/main/java/org/blacksmith/finlib/interestbasis/daycount/ActActYearConvention.java
@@ -8,15 +8,15 @@ public class ActActYearConvention implements DayCountConventionCalculator {
 
   @Override
   public double calculateYearFraction(LocalDate startDate, LocalDate endDate, ScheduleInfo scheduleInfo) {
-    final int years = DateUtils.yearsBetween(startDate, endDate);
+    final long years = DateUtils.yearsBetween(startDate, endDate);
     final LocalDate remainedPeriodStartDate = years==0 ? startDate : startDate.plusYears(years);
-    final int remainedPeriodDays = DateUtils.daysBetween(remainedPeriodStartDate, endDate);
+    final long remainedPeriodDays = DateUtils.daysBetween(remainedPeriodStartDate, endDate);
     final double remainedPeriodYearLength = DateUtils.daysBetween(remainedPeriodStartDate, remainedPeriodStartDate.plusYears(1));
     return years + (remainedPeriodDays/remainedPeriodYearLength);
   }
 
   @Override
-  public int calculateDays(LocalDate startDate, LocalDate calcDate, ScheduleInfo scheduleInfo) {
+  public long calculateDays(LocalDate startDate, LocalDate calcDate, ScheduleInfo scheduleInfo) {
     return DateUtils.daysBetween(startDate, calcDate);
   }
 }
